chore: refactored code so Add Reservation and Edit Reservation pages are combined
ref: sos-1121
